Cleanup: Removal of challenge content
* Moved autoagents out of challenge folder
* Moved scenarios and routes to data folder
* Removed challenge files (Docker and statistics missing)
* Removed utils folder as its files were duplicated
* Removed statistics_manager and challenge_mode
* Removed challenge sensors
* Changed downsampling to match the one at leaderboard
* Removed Track class
* Removed testing and team_code
