--- srunner/challenge/autoagents/agent_wrapper.py
@@ -1,174 +1,0 @@
-#!/usr/bin/env python
-
-# Copyright (c) 2019 Intel Corporation
-#
-# This work is licensed under the terms of the MIT license.
-# For a copy, see <https://opensource.org/licenses/MIT>.
-
-"""
-Wrapper for autonomous agents required for tracking and checking of used sensors
-"""
-
-from __future__ import print_function
-import math
-import os
-
-import carla
-
-from srunner.challenge.envs.scene_layout_sensors import SceneLayoutReader, ObjectFinder
-from srunner.challenge.envs.sensor_interface import CallBack, CANBusSensor, HDMapReader
-from srunner.challenge.autoagents.autonomous_agent import Track
-from srunner.scenariomanager.carla_data_provider import CarlaDataProvider
-
-MAX_ALLOWED_RADIUS_SENSOR = 5.0
-
-
-class SensorConfigurationInvalid(Exception):
-
-    """Base class for other exceptions"""
-
-    def __init__(self, message):
-        print(message)
-        super(SensorConfigurationInvalid, self).__init__()
-
-
-class AgentWrapper(object):
-
-    """
-    Wrapper for autonomous agents required for tracking and checking of used sensors
-    """
-
-    _agent = None
-    _sensors_list = []
-    _challenge_mode = False
-
-    def __init__(self, agent, challenge_mode):
-        """
-        Set the autonomous agent
-        """
-        self._agent = agent
-        self._challenge_mode = challenge_mode
-
-    def __call__(self):
-        """
-        Pass the call directly to the agent
-        """
-        return self._agent()
-
-    def setup_sensors(self, vehicle, debug_mode=False):
-        """
-        Create the sensors defined by the user and attach them to the ego-vehicle
-        :param vehicle: ego vehicle
-        :return:
-        """
-        bp_library = CarlaDataProvider.get_world().get_blueprint_library()
-        for sensor_spec in self._agent.sensors():
-            # These are the pseudosensors (not spawned)
-            if sensor_spec['type'].startswith('sensor.scene_layout'):
-                # Static sensor that gives you the entire information from the world (Just runs once)
-                sensor = SceneLayoutReader(CarlaDataProvider.get_world())
-            elif sensor_spec['type'].startswith('sensor.object_finder'):
-                # This sensor returns the position of the dynamic objects in the scene.
-                sensor = ObjectFinder(CarlaDataProvider.get_world(), sensor_spec['reading_frequency'])
-            elif sensor_spec['type'].startswith('sensor.can_bus'):
-                # The speedometer pseudo sensor is created directly here
-                sensor = CANBusSensor(vehicle, sensor_spec['reading_frequency'])
-            elif sensor_spec['type'].startswith('sensor.hd_map'):
-                # The HDMap pseudo sensor is created directly here
-                sensor = HDMapReader(vehicle, sensor_spec['reading_frequency'])
-            # These are the sensors spawned on the carla world
-            else:
-                bp = bp_library.find(str(sensor_spec['type']))
-                if sensor_spec['type'].startswith('sensor.camera'):
-                    bp.set_attribute('image_size_x', str(sensor_spec['width']))
-                    bp.set_attribute('image_size_y', str(sensor_spec['height']))
-                    bp.set_attribute('fov', str(sensor_spec['fov']))
-                    sensor_location = carla.Location(x=sensor_spec['x'], y=sensor_spec['y'],
-                                                     z=sensor_spec['z'])
-                    sensor_rotation = carla.Rotation(pitch=sensor_spec['pitch'],
-                                                     roll=sensor_spec['roll'],
-                                                     yaw=sensor_spec['yaw'])
-                elif sensor_spec['type'].startswith('sensor.lidar'):
-                    bp.set_attribute('range', str(sensor_spec['range']))
-                    bp.set_attribute('rotation_frequency', str(sensor_spec['rotation_frequency']))
-                    bp.set_attribute('channels', str(sensor_spec['channels']))
-                    bp.set_attribute('upper_fov', str(sensor_spec['upper_fov']))
-                    bp.set_attribute('lower_fov', str(sensor_spec['lower_fov']))
-                    bp.set_attribute('points_per_second', str(sensor_spec['points_per_second']))
-                    sensor_location = carla.Location(x=sensor_spec['x'], y=sensor_spec['y'],
-                                                     z=sensor_spec['z'])
-                    sensor_rotation = carla.Rotation(pitch=sensor_spec['pitch'],
-                                                     roll=sensor_spec['roll'],
-                                                     yaw=sensor_spec['yaw'])
-                elif sensor_spec['type'].startswith('sensor.other.gnss'):
-                    sensor_location = carla.Location(x=sensor_spec['x'], y=sensor_spec['y'],
-                                                     z=sensor_spec['z'])
-                    sensor_rotation = carla.Rotation()
-
-                # create sensor
-                sensor_transform = carla.Transform(sensor_location, sensor_rotation)
-                sensor = CarlaDataProvider.get_world().spawn_actor(bp, sensor_transform, vehicle)
-            # setup callback
-            sensor.listen(CallBack(sensor_spec['id'], sensor, self._agent.sensor_interface))
-            self._sensors_list.append(sensor)
-
-        self._validate_sensor_configuration()
-
-        while not self._agent.all_sensors_ready():
-            if debug_mode:
-                print(" waiting for one data reading from sensors...")
-            CarlaDataProvider.perform_carla_tick()
-
-    def _validate_sensor_configuration(self):
-        """
-        Ensure that the sensor configuration is valid, in case the challenge mode is used
-        Returns true on valid configuration, false otherwise
-        """
-
-        if not self._challenge_mode:
-            return
-
-        phase_codename = os.getenv('CHALLENGE_PHASE_CODENAME', 'dev_track_3')
-        track = int(phase_codename.split("_")[2])
-        phase = phase_codename.split("_")[0]
-
-        if phase != 'debug' and Track(track) != self._agent.track:
-            raise SensorConfigurationInvalid("You are submitting to the wrong track [{}]!".format(Track(track)))
-
-        for sensor in self._agent.sensors():
-            if self._agent.track == Track.ALL_SENSORS:
-                if sensor['type'].startswith('sensor.scene_layout') or sensor['type'].startswith(
-                        'sensor.object_finder') or sensor['type'].startswith('sensor.hd_map'):
-                    raise SensorConfigurationInvalid("Illegal sensor used for Track [{}]!".format(self._agent.track))
-
-            elif self._agent.track == Track.CAMERAS:
-                if not (sensor['type'].startswith('sensor.camera.rgb') or sensor['type'].startswith(
-                        'sensor.other.gnss') or sensor['type'].startswith('sensor.can_bus')):
-                    raise SensorConfigurationInvalid("Illegal sensor used for Track [{}]!".format(self._agent.track))
-
-            elif self._agent.track == Track.ALL_SENSORS_HDMAP_WAYPOINTS:
-                if (sensor['type'].startswith('sensor.scene_layout') or
-                        sensor['type'].startswith('sensor.object_finder')):
-                    raise SensorConfigurationInvalid("Illegal sensor used for Track [{}]!".format(self._agent.track))
-            else:
-                if not (sensor['type'].startswith('sensor.scene_layout') or sensor['type'].startswith(
-                        'sensor.object_finder') or sensor['type'].startswith('sensor.other.gnss')
-                        or sensor['type'].startswith('sensor.can_bus')):
-                    raise SensorConfigurationInvalid("Illegal sensor used for Track [{}]!".format(self._agent.track))
-
-            # let's check the extrinsics of the sensor
-            if 'x' in sensor and 'y' in sensor and 'z' in sensor:
-                if math.sqrt(sensor['x']**2 + sensor['y']**2 + sensor['z']**2) > MAX_ALLOWED_RADIUS_SENSOR:
-                    raise SensorConfigurationInvalid(
-                        "Illegal sensor extrinsics used for Track [{}]!".format(self._agent.track))
-
-    def cleanup(self):
-        """
-        Remove and destroy all sensors
-        """
-        for i, _ in enumerate(self._sensors_list):
-            if self._sensors_list[i] is not None:
-                self._sensors_list[i].stop()
-                self._sensors_list[i].destroy()
-                self._sensors_list[i] = None
-        self._sensors_list = []
