--- srunner/challenge/no_scenarios.json
@@ -1,6 +1,0 @@
-{
-"available_scenarios": [
-{
-}
-]
-}
