+++ srunner/utilities/code_check_and_formatting.sh
@@ -4,16 +4,14 @@ autopep8 scenario_runner.py --in-place --max-line-length=120
 autopep8 srunner/scenariomanager/*.py --in-place --max-line-length=120
 autopep8 srunner/scenariomanager/scenarioatomics/*.py --in-place --max-line-length=120
 autopep8 srunner/scenarios/*.py --in-place --max-line-length=120
-autopep8 srunner/challenge/*.py --in-place --max-line-length=120
-autopep8 srunner/challenge/autoagents/*.py --in-place --max-line-length=120
+autopep8 srunner/autoagents/*.py --in-place --max-line-length=120
 autopep8 srunner/tools/*.py --in-place --max-line-length=120
 autopep8 srunner/scenarioconfigs/*.py --in-place --max-line-length=120
 
 
 pylint --rcfile=.pylintrc --disable=I srunner/scenariomanager/
 pylint --rcfile=.pylintrc srunner/scenarios/
-pylint --rcfile=.pylintrc srunner/challenge/autoagents/
-pylint --rcfile=.pylintrc srunner/challenge/challenge_statistics_manager.py
+pylint --rcfile=.pylintrc srunner/autoagents/
 pylint --rcfile=.pylintrc srunner/tools/
 pylint --rcfile=.pylintrc srunner/scenarioconfigs/
 pylint --rcfile=.pylintrc scenario_runner.py
