--- team_code/README.md
@@ -1,1 +1,0 @@
-Team code goes in this folder
